requirements.txt
streamlit
PyPDF2
python-docx
sentence-transformers
fuzzywuzzy
python-Levenshtein
pandas
